gradescope/requirements.txt
gspread
fullGSapi==1.3.11
python-dotenv==1.0.1
google-api-python-client
google-auth
google-auth-oauthlib
backoff
pandas
backoff_utils
requests
croniter==6.0.0
beautifulsoup4